requirements.txt
python-dotenv
langchain[anthropic]
langchain-community
langchain-google-vertexai
langchain-core
